feat(spurd): converge node device inventory on out-of-band changes (#898)
Closes #800. spurd discovered device inventory once at startup and
heartbeats carried no resources, so an out-of-band GPU change (primarily
an AMD MI300X SPX/CPX compute-partition switch) never reached the
controller, which kept scheduling against hardware that no longer
existed. Add periodic re-discovery that converges the controller without
a spurd restart, and rework GPU allocation accounting to key on a stable
device identity so a mid-set device change cannot corrupt surviving
allocations.

- Stable identity keyed on the PCIe BDF (matching how ROCm's
  k8s-device-plugin and DRA driver surface devices), plus a per-partition
  rank for CPX. Encoded into a u64 stable_id:
  (domain<<24)|(bus<<16)|(dev<<11)|(func<<8)|partition. Carries the PCI
  domain so it does not collide on multi-segment hosts and decodes to the
  DRA pciBusID form (0000:19:00.0). device_id stays the positional
  visible index used for GPU injection (ROCR_VISIBLE_DEVICES).
- Diff-and-classify refresh: a periodic agent task re-discovers
  inventory, converges the free pool, and re-registers so the controller
  converges. Baseline and the shared registry commit only after the
  controller acknowledges.
- An out-of-band change to an allocated GPU is not an inventory event:
  held devices are pinned (total_resources stays a superset of the
  allocation) and fresh partitions overlapping a held device's BDF are
  suppressed, so a second job cannot be double-booked onto held silicon.
- Non-disruptive upgrade: a legacy job's persisted positional GPU ids are
  translated through the current inventory at restore, and the agent
  reports its translated held set so the controller converges
  alloc_resources into stable-id space (no dispatch churn) instead of
  reading busy GPUs as free.
- Generation guard: the controller stamps each allocation with the node
  inventory generation; the agent rejects a stale dispatch under the
  allocation lock. stable_id is backfilled from device_id at every
  controller ingress boundary so legacy inventories do not collide on 0.

All wire changes are additive (no proto renumber; the unreleased
stable_id / allocated-device id fields widen to u64); new serde fields
default so a new controller replays old Raft entries cleanly.

Follow-ups tracked separately: controller-side drain/fail-requeue for a
device physically lost under a running job (#892), and auth-mode
re-registration hardening (#891).

+++ crates/spur-core/src/hooks.rs
@@ -26,7 +26,7 @@ pub struct HookContext {
     /// `prolog_slurmd`, `epilog_slurmd`, `prolog_slurmctld`, `epilog_slurmctld`,
     /// `prolog_task`, `epilog_task`, `prolog_srun`, `epilog_srun`.
     pub script_context: String,
-    pub gpu_devices: Vec<u32>,
+    pub gpu_devices: Vec<u64>,
     pub cpus: u32,
     pub memory_mb: u64,
 }
